server/requirements.txt

# 1. Core API Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.2.1

# 2. Frontend UI
streamlit>=1.32.0

# 3. Database & Storage (100% Local, không cần Server)
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0     # SQL nhẹ nhàng lưu lịch sử chat
chromadb>=0.4.22      # Vector DB (Chế độ lưu file local)
lightrag-hku>=1.0.0   # Knowledge Graph

# 4. Document Parsing 
pymupdf>=1.23.0       # Đọc PDF tốc độ bàn thờ bằng CPU (Core chính)
docling>=2.0.0        # Xử lý bảng biểu khó (Chỉ định chạy CPU)
python-magic>=0.4.27  # Nhận diện định dạng file 

# 5. Core AI & RAG Pipeline 
sentence-transformers>=2.2.2  # Nhúng Vector chạy mượt trên local
flashrank>=0.2.0              # Chấm điểm Reranking cho chuẩn xác
langchain-text-splitters>=0.0.1
langchain-ollama      # LangChain wrapper: Ollama LLM & Embeddings
langchain-google-genai>=2.1.0 # LangChain wrapper: Google Gemini LLM & Embeddings
tiktoken>=0.5.0
google-genai>=1.0.0           # API Gemini 
ollama>=0.4.0                 # Local LLM (Dành trọn 6GB VRAM cho thằng này)

# 6. Observability (Giám sát, check lỗi)
ragas>=0.1.0         # Đo lường AI trả lời ngu hay khôn
arize-phoenix>=4.0.0 # Bắt lỗi đường đi của Data (rất cần cho dev)
openinference-instrumentation-langchain

# 7. Utilities 
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0